created ReferencesTest
